refactor(editor): Move hasProxyAuth into existing nodeTypesUtils.ts
