--- packages/frontend/editor-ui/src/app/utils/nodeTypeUtils.ts
@@ -1,9 +1,0 @@
-import type { INodeUi } from '@/Interface';
-
-/**
- * True when the node uses the HTTP-request proxy-auth pattern
- * (parameter key `nodeCredentialType` is present).
- */
-export function hasProxyAuth(node: INodeUi): boolean {
-	return Object.keys(node.parameters).includes('nodeCredentialType');
-}
